refactor: separate data
