+++ src/lib/data/country.ts
@@ -1,0 +1,11 @@
+export const DEFAULT_FILE_NAME: string = 'file'
+export const COUNTRIES: { [key: string]: string } = {
+  us: 'English',
+  kr: 'Korean',
+  in: 'Indian',
+  ch: 'Chinese',
+  vn: 'Vietnamese',
+  jp: 'Japanese',
+};
+
+export const COUNTRY_CODES = Object.keys(COUNTRIES)
